reFact(NumberGame) : create numberGame package and move interfaces to the package
