+++ src/main/java/lotto/numberGame/JudgeMachine.java
@@ -1,0 +1,5 @@
+package lotto.numberGame;
+
+public interface JudgeMachine extends Machine {
+	int rank(Integer winPoint);
+}
