+++ src/main/java/lotto/numberGame/Machine.java
@@ -1,4 +1,4 @@
-package machines.interfaces;
+package lotto.numberGame;
 
 public interface Machine {
 	void run();
