+++ src/main/java/lotto/numberGame/Validator.java
@@ -1,6 +1,7 @@
-package validate;
+package lotto.numberGame;
 
 public interface Validator {
 	void handleException(boolean condition, String ErrorMessage);
+
 	void printErrorMessage(IllegalArgumentException e);
 }
